chore: change of types

+++ clients/platform/v0/objective-c/Platform.pbobjc.h
@@ -80,7 +80,7 @@ typedef GPB_ENUM(ResponseMetadata_FieldNumber) {
 
 @property(nonatomic, readwrite) int64_t height;
 
-@property(nonatomic, readwrite) int64_t coreChainLockedHeight;
+@property(nonatomic, readwrite) uint32_t coreChainLockedHeight;
 
 @end
